fix converting to QImage from images whose width is not 32-bit aligned
QImage has a requirement that all scanlines must be aligned on 32bit
boundaries. Since we don't require that from our inputs, we need
to ensure that it is done right.

This commit fixes the RGB and GRAYSCALE(1) cases. UYVY has been modified
as well, but is not tested. GRAYSCALE(8) and BAYER are not changed.